persist game & board

+++ app/src/main/java/battleships/esa/ffhs/ch/refactored/data/player/Player.kt
@@ -11,6 +11,6 @@ data class Player(
 ) {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(index = true)
-    var id: Int = 0
+    var id: Long = 0
 }
